errand/requirements.txt
Django==5.2.7
djangorestframework>=3.14
djangorestframework-simplejwt>=5.0
python-dotenv>=1.0.0
requests>=2.31
# Optional (recommended if frontend will call API from browser):
# django-cors-headers>=4.0
